% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ABIS_markers}
\alias{ABIS_markers}
\title{Gene markers of 29 immune cell types}
\format{
dataframe
}
\source{
\href{https://github.com/giannimonaco/ABIS}{ABIS}
}
\usage{
ABIS_markers
}
\description{
List of immune markers from Monaco et al.
}
\examples{
data(ABIS_markers)
head(ABIS_markers)
unique(ABIS_markers$Cell.type)
}
\references{
Monaco et al. (2019) Cell reports vol. 26,6
(\href{https://pubmed.ncbi.nlm.nih.gov/30726743/}{PubMed})
}
\keyword{datasets}
