% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{acc_sim}
\alias{acc_sim}
\title{Utility function to generate accuracy metrics, for use with \code{\link[=estimate_accuracy]{estimate_accuracy()}}}
\usage{
acc_sim(n, method = "Uniform", p = NULL, ...)
}
\arguments{
\item{n}{An integer giving the desired sample size for which the target function is to be calculated.}

\item{method}{An optional string stating the distribution from which data is to be generated. Default is i.i.d. uniform sampling. Currently also supports "Class Imbalance". Can also take a function outputting a vector of probabilities if the user wishes to specify a custom distribution.}

\item{p}{If method is 'Class Imbalance', gives the degree of weight placed on the positive class.}

\item{...}{Additional model parameters to be specified by the user.}
}
\value{
A data frame giving performance metrics for the specified sample size.
}
\description{
Utility function to generate accuracy metrics, for use with \code{\link[=estimate_accuracy]{estimate_accuracy()}}
}
