% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{add_adt_assay}
\alias{add_adt_assay}
\title{Add ADT information to a seurat object}
\usage{
add_adt_assay(srtObject, exp_metadata, dataDIR, outs = FALSE)
}
\arguments{
\item{srtObject}{Seurat object}

\item{exp_metadata}{Metadata table}

\item{dataDIR}{Directory containing the cellranger output directory}

\item{outs}{Outs dir or not | Default \code{FALSE}}
}
\description{
Add ADT information to a seurat object
}
\examples{
\dontrun{add_adt_assay()}
}
\keyword{seurat}
