% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{add_vdj_assay}
\alias{add_vdj_assay}
\title{Add VDJ information to a seurat object}
\usage{
add_vdj_assay(srtObject, exp_metadata, dataDIR)
}
\arguments{
\item{isVerbose}{Show all the messages to screen | Default \code{FALSE}}
}
\description{
Add VDJ information to a seurat object
}
\examples{
\dontrun{add_vdj_assay()}
}
\keyword{seurat}
