% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcR_functions.R
\name{archR_subset_by_sample}
\alias{archR_subset_by_sample}
\title{Subset an arcR projects based on sampleID value}
\usage{
archR_subset_by_sample(arcR_project, ident_value)
}
\arguments{
\item{arcR_project}{Directory in which the project will be located.  Default is "getwd()"}

\item{ident_value}{Name of the project}
}
\value{
An archR project
}
\description{
Given a sample name, the function returns an archR project containing only that 
sample.
}
\examples{
\dontrun{
# If using built-in directory structure.
new_project <- archR_subset_by_sample(exp_Integrated,"C1")
}

}
\concept{ATAC-Seq}
