% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdj_functions.R
\name{bcr_lineage_reconstruction}
\alias{bcr_lineage_reconstruction}
\title{Build and visualise phylogenetic trees on AIRR formatted file}
\usage{
bcr_lineage_reconstruction(
  germ_file,
  clonesID = NULL,
  method = "pratchet",
  exec = NULL,
  scale = 0.01
)
}
\arguments{
\item{germ_file}{AIRR formatted table}

\item{clonesID}{List of clones to build the tree for. If \code{NULL} all the clones will be used | Default \code{NULL}}

\item{method}{Method used to estimate tree topology and branch lengths. Available methods are: pratchet, pml, dnapars, dnaml and igphyml | Default \code{pratchet}}

\item{exec}{Path to dnapars executable (only used if \code{method = dnapars}) | Default \code{NULL}}

\item{scale}{With of the scale bar | Default \code{0.01}}
}
\description{
Build and visualise phylogenetic trees on AIRR formatted file
}
\examples{
\dontrun{
bcr_lineage_reconstruction("IGHV-genotyped_M2_germ-pass.tab",c(379,375,455,291),scale = 1)
}
}
