% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{br}
\alias{br}
\title{Replication data for 'Predicting Recidivism'}
\format{
An object of class \code{data.frame} with 7214 rows and 14 columns.
}
\usage{
br
}
\description{
Replication data for 'Predicting Recidivism'
}
\references{
\url{https://www.science.org/doi/full/10.1126/sciadv.aao5580}
}
\author{
Julia Dressel and Hany Farid
}
\keyword{datasets}
