% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{calculate_cell_complexity}
\alias{calculate_cell_complexity}
\title{Calculate the complexity of the dataset}
\usage{
calculate_cell_complexity(seurat_object)
}
\arguments{
\item{seurat_object}{seurat object}
}
\description{
Calculate the complexity of the dataset
}
\details{
Details
Refer to the pre-processing article for a detailed usage and examples
}
\examples{
\dontrun{
calcualte_cell_complexity(seurat_object)
}
}
\keyword{seurat}
