% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{calculate_cell_proportions}
\alias{calculate_cell_proportions}
\title{Calculate cell proportions}
\usage{
calculate_cell_proportions(
  srtObject,
  cluster_col,
  groups_annotation,
  mycolors,
  plotType,
  plotTitle,
  sampleOrder = NULL,
  groupOrder = NULL,
  noClusters = FALSE,
  stats_compare = FALSE,
  png = TRUE
)
}
\arguments{
\item{srtObject}{Seurat object.}

\item{cluster_col}{Metadata column containing the clustering information.}

\item{groups_annotation}{data.frame containing sample information.}

\item{mycolors}{something}

\item{plotType}{something}

\item{plotTitle}{something}

\item{sampleOrder}{something}

\item{groupOrder}{something}

\item{noClusters}{something}

\item{stats_compare}{something}

\item{png}{something}
}
\description{
This function calculate the cell proportions.
}
\examples{
\dontrun{calculate_cell_proportions()}

}
\keyword{seurat}
