% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_utils.R
\name{calculate_genes_signature}
\alias{calculate_genes_signature}
\title{Calcualte genes signature}
\usage{
calculate_genes_signature(srtObject, markers_list, assay = "RNA", ncores = 2)
}
\arguments{
\item{srtObject}{Seurat object.}
}
\description{
This function
}
\examples{
\dontrun{
calcualte_genes_signature(exp_Integrated, ".", "integrated_snn_res.0.8")
}
}
\keyword{markers}
\keyword{seurat}
