% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{calculate_mito_ribo}
\alias{calculate_mito_ribo}
\title{Calculate the percentage of mitochondrial and ribosomal genes}
\usage{
calculate_mito_ribo(
  seurat_object,
  organism = "human",
  new_mito = NULL,
  new_ribo = NULL
)
}
\arguments{
\item{seurat_object}{seurat object}

\item{organism}{Human by default.}

\item{new_mito}{mitochondrial pattern to use if the organisms is not human or mouse}

\item{new_ribo}{ribosomial pattern to use if the organisms is not human or mouse}
}
\description{
Calculate the percentage of mitochondrial and ribosomal genes
}
\details{
Details
Refer to the pre-processing article for a detailed usage and examples
}
\examples{
\dontrun{
calculate_mito_ribo(seurat_object)
}
}
\keyword{seurat}
