% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{CellCycleScoring_extra}
\alias{CellCycleScoring_extra}
\title{Expansion of CellCycleScoring from the seurat package.}
\usage{
CellCycleScoring_extra(
  object,
  s.features,
  g2m.features,
  g1s.features,
  m.features,
  mg1.features,
  ctrl = NULL,
  set.ident = FALSE,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{s.features}{List of genes in the S phase}

\item{g2m.features}{List of genes in the G2M phase}

\item{g1s.features}{List of genes in the G1S phase}

\item{m.features}{List of genes in the M phase}

\item{mg1.features}{List of genes in the MG1 phase}

\item{ctrl}{Number of control features selected from the same bin per analyzed feature supplied to AddModuleScore. Defaults to value equivalent to minimum number of features present in 's.features' and 'g2m.features'.}

\item{set.ident}{If true, sets identity to phase assignments Stashes old identities in 'old.ident'}

\item{...}{Arguments to be passed to AddModuleScore}
}
\description{
In addition to the default phases, CellCycleScoring_extra supports 2 more cell phases.
}
\examples{
\dontrun{CellCycleScoring_extra()}
}
\keyword{seurat}
