% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{compare_qc}
\alias{compare_qc}
\title{Compare 2 seurat object (used to generate QC figures between pre-fitlered and filtered data)}
\usage{
compare_qc(
  object1,
  object2,
  comp_name,
  colori = NULL,
  qc_dir = ".",
  minFeatures = 500,
  mtPerc = 20
)
}
\arguments{
\item{object1}{seurat object 1}

\item{object2}{seurat object 2}

\item{comp_name}{name to use for the figures}

\item{colori}{set of 2 colors to use for plotting | Default \code{c("#D3D3D3","#FFB140")}}

\item{qc_dir}{output directory}

\item{minFeatures}{number of genes}

\item{mtPerc}{percentage fo mt genes}
}
\description{
Compare 2 seurat object (used to generate QC figures between pre-fitlered and filtered data)
}
\details{
Details
Refer to the pre-processing article for a detailed usage and examples
}
\examples{
\dontrun{
compare_qc(object1, object2, "comparison_1")
}
}
\keyword{seurat}
