% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scR.R
\name{conduct_interpolation}
\alias{conduct_interpolation}
\title{Conduct interpolation on a single simulation}
\usage{
conduct_interpolation(
  scbobject,
  epsilon,
  delta,
  maxN,
  delta_formula,
  epsilon_formula,
  delta_lower_bounds = NULL,
  epsilon_lower_bounds = NULL,
  delta_upper_bounds = NULL,
  epsilon_upper_bounds = NULL,
  delta_start = NULL,
  epsilon_start = NULL
)
}
\arguments{
\item{scbobject}{An object of class "scb_data" for interpolation to be conducted on.}

\item{epsilon}{A real number between 0 and 1 giving the targeted maximum out-of-sample (OOS) error rate}

\item{delta}{A real number between 0 and 1 giving the targeted maximum probability of observing an OOS error rate higher than \code{epsilon}}

\item{maxN}{A positive integer giving value of the largest N for which extrapolation is to be conducted.}

\item{delta_formula}{Formula of the form Delta ~ model(n,...) giving the NLS model to be applied to the delta curve.}

\item{epsilon_formula}{Formula of the form Epsilon ~ model(n,...) giving the NLS model to be applied to the epsilon curve.}

\item{delta_lower_bounds}{Optional named vector of upper bounds for the model parameters.}

\item{epsilon_lower_bounds}{A formula object giving the model to be fit.}

\item{delta_upper_bounds}{A data frame giving the data the model is to be fit on.}

\item{epsilon_upper_bounds}{An integer vector of N values to conduct interpolation and extrapolation on.}

\item{delta_start}{A positive integer giving value of the largest N in the observed data.}

\item{epsilon_start}{Optional named vector of starting values for the model parameters.}
}
\value{
A named \code{list} containing the interpolated dataframe, the original input data frame, and the given values of epsilon, delta, and maxN.
}
\description{
Wrapper function for fitting extrapolation model to a single object of class "scb_data". Allows fitting of custom functions, but for general use interpolate_scb should be used instead.
}
\seealso{
\code{\link[=interpolate_scb]{interpolate_scb()}} is the main wrapper for interpolation on a list.
}
