% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert_ensemble_to_symbol}
\alias{convert_ensemble_to_symbol}
\title{Convert the rownames of a matrix from ensemble to symbol}
\usage{
convert_ensemble_to_symbol(matrice)
}
\arguments{
\item{genelist}{a matrix with rownames in ensemle format}
}
\description{
Convert the rownames of a matrix from ensemble to symbol
}
\examples{
\dontrun{convert_ensemble_to_symbol(matrice)}
}
