% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scR.R
\name{create_scb_model}
\alias{create_scb_model}
\title{Create custom model fitting function}
\usage{
create_scb_model(
  model_fun,
  extra_args = list(),
  split = FALSE,
  arg_map = list(formula = "formula", data = "data", x = "x", y = "y")
)
}
\arguments{
\item{model_fun}{A binary classification model supplied by the user, e.g. glm}

\item{extra_args}{A list of additional default arguments to be passed to the classification model, e.g. family = binomial(link = "logit")}

\item{split}{Logical indicating whether the model expects a single data argument or separate x/y values.}

\item{arg_map}{Named list giving mappings for names of formula, data, x, and y arguments expected in \code{model_fun}. Must include formula and data if split is FALSE or x and y otherwise.}
}
\value{
A function taking supplied arguments that can be passed to other package functions such as \code{\link[=estimate_accuracy]{estimate_accuracy()}}
}
\description{
Utility function for creating custom classification function for use in SCB calculations.
}
