% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scR.R
\name{create_scb_prediction}
\alias{create_scb_prediction}
\title{Utility function for creating custom prediction function for use in SCB calculations.}
\usage{
create_scb_prediction(
  predict_fun,
  extra_args = list(),
  transform_fn = identity,
  arg_map = list(m = "m", newdata = "newdata")
)
}
\arguments{
\item{predict_fun}{A binary prediction model supplied by the user, e.g. predict.glm}

\item{extra_args}{A list of additional default arguments to be passed to the classification model, e.g. type="response"}

\item{transform_fn}{Function giving the transformation, if any, to be applied to the prediction output to generate binary predictions. Defaults to identity()}

\item{arg_map}{Named list giving mappings for names of \code{m} and \code{newdata} arguments expected in \code{predict_fun}.}
}
\value{
A prediction function taking supplied arguments that can be passed to other package functions such as \code{\link[=estimate_accuracy]{estimate_accuracy()}}
}
\description{
Utility function for creating custom prediction function for use in SCB calculations.
}
