% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dendrite_ex}
\alias{dendrite_ex}
\title{Example of the dendrite data}
\format{
\subsection{\code{dendrite_ex}}{

A data frame with 52 rows and 2 columns.
These are xy-coordinates of the data.
}
}
\usage{
dendrite_ex
}
\description{
An Example of the dendrite data for 2d plot. This is converted from dendrite data in the library "spatstat".
}
\keyword{datasets}
