% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions.R
\name{enrichR_analysis}
\alias{enrichR_analysis}
\title{Run enrichR}
\usage{
enrichR_analysis(
  genes2enrich,
  databases = NULL,
  n_top = NULL,
  direction = "up",
  out_dir = "."
)
}
\arguments{
\item{genes2enrich}{.}

\item{databases}{.}

\item{n_top}{.}

\item{direction}{.}

\item{out_dir}{.}
}
\description{
Run enrichR
}
\examples{
\donttest{

}
}
\keyword{enrichR}
