% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{estimate_ambient_rna}
\alias{estimate_ambient_rna}
\title{Estimate ambient RNA contamination using SoupX}
\usage{
estimate_ambient_rna(
  sampleid,
  raw_dataDIR,
  estGenes = NULL,
  counts2disk = FALSE
)
}
\arguments{
\item{sampleid}{Sample ID}

\item{raw_dataDIR}{Directory containing the cellranger outputs (e.g. /outs)}

\item{estGenes}{List of genes to use to estimate the contamination | Default \code{NULL}}

\item{counts2disk}{Save the corrected counts as h5 | Default \code{FALSE}}
}
\description{
Estimate ambient RNA contamination using SoupX
}
\details{
Details
Refer to the pre-processing article for a detailed usage and examples
}
\examples{
\dontrun{
estimate_ambient_rna("Sample1",".", estGenes = c("PPBP","HBB","HBA1"), FALSE)
}
}
\keyword{soupX}
