% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{export_seurat}
\alias{export_seurat}
\title{Export a seurat object to file}
\usage{
export_seurat(srtObject, Assay = "RNA", Slot = "data", bpcells = FALSE)
}
\arguments{
\item{srtObject}{seurat object to export}
}
\description{
Convenient wrapper to export a seurat object to files
}
\examples{
\dontrun{
export_seurat(pbmc_small)
}
}
