% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions.R
\name{fgsea_rank_genes}
\alias{fgsea_rank_genes}
\title{Rank genes for FGSEA analysis}
\usage{
fgsea_rank_genes(deg_output, isHuman = TRUE, entrez = TRUE)
}
\arguments{
\item{deg_output}{.}

\item{isHuman}{.}

\item{entrez}{.}
}
\description{
Rank genes for FGSEA analysis
}
\examples{
\donttest{

}
}
\keyword{FGSEA}
