% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scR.R
\name{fit_and_predict}
\alias{fit_and_predict}
\title{Utility function to fit extrapolation model, for use with \code{\link[=conduct_interpolation]{conduct_interpolation()}}}
\usage{
fit_and_predict(
  formula,
  data,
  N_grid,
  maxN_obs,
  start = NULL,
  lower = NULL,
  upper = NULL
)
}
\arguments{
\item{formula}{A formula object giving the model to be fit.}

\item{data}{A data frame giving the data the model is to be fit on.}

\item{N_grid}{An integer vector of N values to conduct interpolation and extrapolation on.}

\item{maxN_obs}{A positive integer giving value of the largest N in the observed data.}

\item{start}{Optional named vector of starting values for the model parameters.}

\item{lower}{Optional named vector of lower bounds for the model parameters.}

\item{upper}{Optional named vector of upper bounds for the model parameters.}
}
\value{
A fitted model object of the chosen type.
}
\description{
Utility function to fit extrapolation model, for use with \code{\link[=conduct_interpolation]{conduct_interpolation()}}
}
