% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{get_loupe_files}
\alias{get_loupe_files}
\title{Dump umap coordinates and metadata to load in the loupe browser}
\usage{
get_loupe_files(srtObject, cellranger_cells, meta_cols, outDIR = NULL)
}
\arguments{
\item{isVerbose}{Show all the messages to screen | Default \code{FALSE}}
}
\description{
Dump umap coordinates and metadata to load in the loupe browser
}
\examples{
\dontrun{get_loupe_files()}
}
\keyword{seurat}
