% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{load_10x_gem}
\alias{load_10x_gem}
\title{Load 10x GEM data}
\usage{
load_10x_gem(
  rawDIR,
  source = "raw",
  source_type = "dir",
  input_matrix = NULL,
  fixed = FALSE
)
}
\arguments{
\item{rawDIR}{Directory containing the data to load}

\item{source}{Use the data from the \code{raw} or \code{filtered} cellranger matrix | Default \code{raw}}

\item{source_type}{| Default \code{dir}}

\item{input_matrix}{}

\item{sample_name}{Name of the sample | Optional}
}
\value{
Returns 
\item{}{a matrix or a list, if multiple type are included}
}
\description{
This function is used to reads the output from cellranger and return the matrices.
}
\examples{
\dontrun{
raw_feature_matrix <- load_10x_gem("/mnt/disk2/raw_files/scanpy_test/Biopsie_17", sample_name = "Biopsie_17", source = "raw", source_type="dir")
}

}
\keyword{seurat}
