% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_utils.R
\name{markers_lookup}
\alias{markers_lookup}
\title{Yes, another parser}
\usage{
markers_lookup(
  object,
  Markers,
  areMarkers = FALSE,
  listCells = FALSE,
  markers_file_path = NULL,
  scale_data = TRUE,
  cluster = FALSE,
  cut = 1,
  ncolumns = NULL,
  colori = c("#FDE725FF", "#440154FF")
)
}
\arguments{
\item{object}{File}

\item{Markers}{FALSE}

\item{areMarkers}{FALSE}

\item{listCells}{FALSE}

\item{markers_file_path}{NULL}

\item{scale_data}{FALSE}

\item{cluster}{FALSE}

\item{cut}{1}

\item{ncolumns}{NULL}

\item{colori}{c("#FAC127FF","#280B54FF")}
}
\description{
Yes, another parser
}
\examples{
\dontrun{
markers_lookup()
}
}
\keyword{markers}
