% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCsim9.R
\name{MCMCsim9}
\alias{MCMCsim9}
\title{MCMCsim9}
\usage{
MCMCsim9(...)
}
\arguments{
\item{burn}{a number indicating the burn-in time. Defaults to 1000.}

\item{m}{a number indicating the length of the generated Markov Chain. Defaults to 15000.}

\item{samole.method}{a charcter string specifying the Metropolis-Hsating method used to
generate the Markov Chain, must be one of "MHS"(default), "IS" or "RWS",
representing Metropois-Hsting sampler, Independent Sampler,
and Random Walk Sampler respectively.}
}
\value{
`estimate theta hat` a number indicating the estimate theta.

method a character string indicating which method is used.

data.name a character string giving the names of the data

x a list of number containing the generated normal chain Xt.

reject_rate a number indicating the reject rate in this method.
}
\description{
The The Metropolis-Hastings algorithms are a class of
Markov Chain Monte Carlo methods including the Metropolis-Hasting sampler,
the Gibbs sampler, the independence sampler, and the random walk.

Corresponding to Rizzo's book exercise 9.6, this function could generated
Markov Chains with any of the Metropolis-Hasting Method, the independence
Metropolis-Hasting Method, and the Random Walk Method for this problem by
specifying the `sample.method` term in the function.
}
\examples{
MCMCsim9(sample.method = "IS", burn = 1000, m = 10000)
MCMCsim9(sample.method = "RWS", burn = 0, m = 10000)
MCMCsim9()
}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
\author{
Tianran Zhang <trzhang@gmail.com>
}
