% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCsimCL1.R
\name{MCsimCL1}
\alias{MCsimCL1}
\title{Monte Carlo experiment to estimate a confidence level}
\usage{
MCsimCL1(...)
}
\arguments{
\item{alpha}{a number indicating significance level. Defaults to 0.05.}

\item{n}{a number indicating random samples size generated from a lognormal distribution. Defaults to 20.}

\item{m}{a number indicating number of response vectors to simulate. Defaults to 1000.}

\item{type}{a character string specifying the parameter to estimate, must be one of "mean"(default),"var".}

\item{interval}{a character string specifying the interval type, must be one of "two.sided"(default) , "right" or "left".}

\item{data}{a character string specifying the data distribution, must be one of "norm"(default),"chisq".}

\item{mu}{a number indicating sample mean of normal distributed data. Defaults to 0.}

\item{sd}{a number indicating sample standard deviation of normal distributed data. Defaults to 1.}

\item{nv}{a number indicating degree freedom of chisq distributed data. Defaults to 1.}
}
\value{
A list containing the following components:

sample estimates:	the value of the estimated confidence level.

method a character string indicating what type interval was simulated.

data.name	a character string giving the name(s) of the data.
}
\description{
Monte Carlo method to assess the confidence level in an estimation procedure.
Empirical confidence level is an estimate of the confidence level obtained by simulation.
For the simulation experiment, repeat the steps above a large number of times,
and compute the proportion of intervals that contain the target parameter.
}
\details{
The formula interface is only applicable for the one sample simulation

If type="norm", parameter mu and sd are used to simulate normal distributed sample. If type="chisq", parameter nv is used to simulate chi-square distributed sample

interval="left" is the interval from one-sided lower confidence bound to positive infinite.
}
\examples{
MCsimCL1(0.01,100,1000,type="mean",interval="two.sided",data="norm")
MCsimCL1(type="mean",interval="two.sided",data="chisq")
}
\author{
Yinuo Liu< smashingzing at gmail.com >
}
