% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Middle_Number}
\alias{Middle_Number}
\title{Calcualte the middle value}
\usage{
Middle_Number(min, max)
}
\arguments{
\item{min}{min value}

\item{max}{max value}
}
\description{
Calcualte the middle value
}
\examples{
\dontrun{Middle_Number(min = 10, max = 20)}
}
