% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_utils.R
\name{modify_scales}
\alias{modify_scales}
\title{Modify scales of a Feature Plot.}
\usage{
modify_scales(p, feature, min.cutoff = NULL, max.cutoff = NULL)
}
\arguments{
\item{p}{Plot resulting from calling SCpubr::do_FeaturePlot() with a single feature and no extra parameters.}

\item{feature}{Feature used in p.}

\item{min.cutoff}{Minimum range of the scale.}

\item{max.cutoff}{Maximum range of the scale.}
}
\value{
A ggplot2 object with the scales modified.
}
\description{
This function transforms all the values in a FeaturePlot that are below min.cutoff to min.cutoff and above max.cutoff to max.cutoff, effectively
achieving the same behaviour as in Seurat::FeaturePlot. All credits for this idea goes to the Seurat developers.
}
\examples{
.
}
