% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MouseHumanGenes}
\alias{MouseHumanGenes}
\title{Mouse-Human gene name mapping table}
\format{
txt
}
\usage{
MouseHumanGenes
}
\description{
Table with gene names in human and mouse format.
}
\examples{
data(MouseHumanGenes)
head(MouseHumanGenes)
}
\keyword{datasets}
