% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_utils.R
\name{parse_input}
\alias{parse_input}
\title{Parser}
\usage{
parse_input(file_str, debug = F)
}
\arguments{
\item{file_str}{File}

\item{debug}{FALSE}
}
\description{
Parser
}
\examples{
\dontrun{
parse_input()
}
}
\keyword{markers}
