% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_utils.R
\name{parse_markers_lists}
\alias{parse_markers_lists}
\title{Another parser}
\usage{
parse_markers_lists(marker_file, useDefault = TRUE)
}
\arguments{
\item{marker_file}{File}

\item{useDefault}{FALSE}
}
\description{
Another parser
}
\examples{
\dontrun{
parse_markers_lists()
}
}
\keyword{markers}
