% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_1k}
\alias{pbmc_1k}
\title{PBMC 1K test dataset}
\format{
seurat
}
\usage{
pbmc_1k
}
\description{
Small pbmc dataset with around 1500 cells.
}
\examples{
data(pbmc_1k)
DimPlot(pbmc_1k, label = TRUE, repel = TRUE) + NoLegend()
}
\keyword{datasets}
