% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_markers}
\alias{pbmc_markers}
\title{PBMC markers}
\format{
garnet marker file
}
\usage{
pbmc_markers
}
\description{
List of markers used to annotate PBMC cell type
}
\examples{
data(pbmc_markers)
}
\keyword{datasets}
