% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_markers_annotated}
\alias{pbmc_markers_annotated}
\title{PBMC markers as data frame}
\format{
dataframe
}
\usage{
pbmc_markers_annotated
}
\description{
PBMC markers
}
\examples{
data(pbmc_markers_annotated)
names(pbmc_markers_annotated)
unique(pbmc_markers_annotated$T)
}
\keyword{datasets}
