% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{PCA_genes}
\alias{PCA_genes}
\title{Function that, given a DESeq2 object, runs PCA and it returns the contributions of the  
genes to each PCs.}
\usage{
PCA_genes(vst_object, ntop = 500, pc_sel = 1, returnPCA = FALSE)
}
\arguments{
\item{vst_object}{vst normalised object (e.g. DESeq2)}

\item{ntop}{Number of genes to use}

\item{pc_sel}{Which PC to use for ordering the result table}

\item{returnPCA}{Return the pca object (FALSE by default)}
}
\description{
Function that, given a DESeq2 object, runs PCA and it returns the contributions of the  
genes to each PCs.
}
\examples{
\dontrun{PCA_genes(vst)}
}
