% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{plot.scb_data}
\alias{plot.scb_data}
\title{Plot method for simulated sample complexity bounds (\code{scb_data} object)}
\usage{
\method{plot}{scb_data}(
  x,
  metrics = c("Accuracy", "Precision", "Recall", "Fscore", "Delta", "Epsilon", "Power"),
  plottype = c("ggplot", "plotly"),
  letters = c("greek", "latin")
)
}
\arguments{
\item{x}{An object of class \code{"scb_data"}, typically the output of \code{\link[=estimate_accuracy]{estimate_accuracy()}}}

\item{metrics}{A character vector containing the metrics to display in the plot. Can include any of "Accuracy", "Precision", "Recall", "Fscore", "Delta", "Epsilon", "Power".}

\item{plottype}{A string indicating the graphics system to use. Must be either \code{"ggplot"} or \code{"plotly"}.}

\item{letters}{A string specifying whether \code{"Delta"} and \code{"Epsilon"} should appear as Greek letters (\code{"greek"}) or Latin (\code{"latin"}) in the legend. Defaults to \code{"greek"}.}
}
\value{
A \link[ggplot2]{ggplot} or \link[plotly]{plot_ly} object, depending on the value of \code{plottype}.
}
\description{
This method plots performance metrics estimated using \code{\link[=estimate_accuracy]{estimate_accuracy()}} for objects of class \code{"scb_data"}.
}
\examples{
\donttest{
}
}
\seealso{
\code{\link[=estimate_accuracy]{estimate_accuracy()}}, which generates the data used for plotting.
}
