% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_utils.R
\name{plot_cells_markers}
\alias{plot_cells_markers}
\title{Plot cell markers}
\usage{
plot_cells_markers(
  srtObject,
  figDIR,
  cluster_col,
  srt_assay = "RNA",
  overlapPlot = NULL
)
}
\arguments{
\item{srtObject}{Seurat object.}

\item{figDIR}{output directory}

\item{cluster_col}{Metadata column containing the clustering information.}

\item{overlapPlot}{something}
}
\description{
This function uses the gene markers defined in the file \code{pbmc_annotation_markers.csv}
to plot and save their expression in both dotplot and featureplot.
}
\examples{
\dontrun{
plot_cells_markers(exp_Integrated, figDIR, "integrated_snn_res.0.8", overlapPlot=NULL)
}
}
\keyword{markers}
\keyword{seurat}
