% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions.R
\name{plot_enrichR_result}
\alias{plot_enrichR_result}
\title{Generate the enrichment figure for enrichR}
\usage{
plot_enrichR_result(enrichR_table, nPaths, colori = NULL, title = "")
}
\arguments{
\item{genes2enrich}{.}

\item{databases}{.}

\item{n_top}{.}

\item{direction}{.}

\item{out_dir}{.}
}
\description{
Generate the enrichment figure for enrichR
}
\examples{
\donttest{

}
\donttest{

}
}
\keyword{enrichR}
