% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scR.R
\name{risk_bounds}
\alias{risk_bounds}
\title{Utility function to generate data points for estimation of the VC Dimension of a user-specified binary classification algorithm given a specified sample size.}
\usage{
risk_bounds(x, ...)
}
\arguments{
\item{x}{An integer giving the desired sample size for which the target function is to be approximated.}

\item{...}{Additional model parameters to be specified by the user.}
}
\value{
A real number giving the estimated value of Xi(n), the bounding function
}
\description{
Utility function to generate data points for estimation of the VC Dimension of a user-specified binary classification algorithm given a specified sample size.
}
