% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{run_doubletFinder}
\alias{run_doubletFinder}
\title{Run DoubletFinder}
\usage{
run_doubletFinder(
  srtObject,
  meta_column = "seurat_clusters",
  isSCT = FALSE,
  countMatrix = FALSE,
  seurat5 = FALSE
)
}
\arguments{
\item{srtObject}{An unfiltered seurat object}

\item{meta_column}{Column metadata containing the cell clustering}

\item{isSCT}{SCT normalisation | Default \code{FALSE}}
}
\value{
Returns 
\item{}{a seurat object}
}
\description{
This function takes a seurat object and calculates doublet/singlet content using DoubletFinder
}
\examples{
\dontrun{
sample_rds <- run_doubletFinder(sample_rds)
}

}
\keyword{seurat}
