% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_pipelines.R
\name{run_integration_with_Ref}
\alias{run_integration_with_Ref}
\title{Samples reference based integration using seurat}
\usage{
run_integration_with_Ref(
  dataDIR,
  pipeline,
  srt_reduction,
  outname,
  ref_samples,
  ref_as_rds = FALSE,
  pcDims = 30,
  isVerbose = FALSE,
  compress_rds = TRUE
)
}
\arguments{
\item{dataDIR}{Path to filtered rds files.}

\item{pipeline}{Use SCT or lognorm normalisation.}

\item{srt_reduction}{Use cca or rpca.}

\item{outname}{Name of the intergated object.}

\item{ref_samples}{List of samples to use as reference.}

\item{ref_as_rds}{Boolean. FALSE by default
\itemize{
\item If TRUE 
\item If FALSE
}}

\item{pcDims}{Number of PCs (30 by default)}

\item{isVerbose}{Show all the messages to screen.FALSE by default}
}
\description{
This function takes a list of rds files (generated by `run_qc_and_filtering`) and integrates them together.
}
\examples{
\dontrun{
run_integration_with_Ref(".", "SCT", "cca", pcDims = 30, "intergation_WRef", c("sample1","sample12"), ref_as_rds = FALSE, isVerbose = FALSE)
}
}
\keyword{seurat}
