% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_pipelines.R
\name{run_seurat_integration}
\alias{run_seurat_integration}
\title{Samples integration using seurat}
\usage{
run_seurat_integration(
  dataDIR,
  pipeline,
  srt_reduction,
  outname,
  nfeatures = 3000,
  pcDims = 30,
  kvalue = 100,
  isVerbose = FALSE,
  vst_flavor = NULL
)
}
\arguments{
\item{dataDIR}{Path to filtered rds files.}

\item{pipeline}{Use SCT or lognorm normalisation.}

\item{srt_reduction}{Use cca or rpca.}

\item{outname}{Name of the intergated object.}

\item{nfeatures}{Number of genes to use when integrating (3000 by default)}

\item{pcDims}{Number of PCs (30 by default)}

\item{kvalue}{Set the number of neighbors used when picking and weighting anchors (100 by default)}

\item{isVerbose}{Show all the messages to screen.FALSE by default}

\item{vst_flavor}{sctransform version}
}
\description{
This function takes a list of rds files (generated by `run_qc_and_filtering`) and integrates them together.
}
\examples{
\dontrun{
run_seurat_integration(".","SCT", "cca", "integration1", 3000, 30, FALSE , "v2")
}
}
\keyword{seurat}
