% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_pipelines.R
\name{run_seurat_integration_v5}
\alias{run_seurat_integration_v5}
\title{Samples integration using seurat v5 (logNorm only)}
\usage{
run_seurat_integration_v5(
  dataDIR,
  samples_vect = NULL,
  intMethod,
  outname,
  nfeatures = 3000,
  pcDims = 30,
  isVerbose = FALSE,
  isSCT = FALSE,
  vst_flavor = NULL,
  ref_samples = NULL,
  merge_layers = TRUE,
  seurat5 = TRUE,
  bpcells_dir = NULL,
  checkpoint = FALSE
)
}
\arguments{
\item{dataDIR}{Path to filtered rds files.}

\item{intMethod}{d}

\item{outname}{Name of the intergated object.}

\item{nfeatures}{Number of genes to use when integrating (3000 by default)}

\item{pcDims}{Number of PCs (30 by default)}

\item{isVerbose}{Show all the messages to screen.FALSE by default}

\item{vst_flavor}{sctransform version}
}
\description{
Samples integration using seurat v5 (logNorm only)
}
\examples{
\dontrun{
run_seurat_integration(".","SCT", "cca", "integration1", 3000, 30, FALSE , "v2")
}
}
\keyword{seurat}
