% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{SaveFigure}
\alias{SaveFigure}
\title{Wrapper function to save figures to disk.}
\usage{
SaveFigure(
  plot_to_save,
  fig_path = ".",
  output_name,
  type = "png",
  width,
  height,
  res = 300,
  ggplot = TRUE
)
}
\arguments{
\item{plot_to_save}{Figure to save}

\item{fig_path}{Where to save the figure}

\item{output_name}{Output file name}

\item{type}{Save as png or pdf (png by default)}

\item{width}{Set the figure width}

\item{height}{Set the figure height}

\item{res}{dpi (300 by default)}

\item{ggplot}{Save the figure using ggsave (TRUE by default)}
}
\description{
Convenient wrapper to save figures with either ggplot (default) or base R.
}
\examples{
\dontrun{
aPlot <- ggplot(...)
SaveFigure(aPlot, "./figures","figure_1","png",1400,900,300)
}
}
