% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCR-package.R
\docType{package}
\name{SCR-package}
\alias{SCR-package}
\title{SCR: Sequential Covariate-Adjusted Randomization}
\description{
Implements the sequential covariate-adjusted randomization (SCR) of
Yang, Qin, Li and Hu (2024), which assigns participants one at a time by
hierarchically minimizing the marginal imbalance and a modified
Mahalanobis distance.
}
\details{
Main functions:
\itemize{
\item \code{\link[=SCR]{SCR()}}: allocate a sequence of units.
\item \code{\link[=scr_next]{scr_next()}}: allocate one newly arrived unit.
\item \code{\link[=scr_imbalance]{scr_imbalance()}}: covariate and marginal imbalance measures.
\item \code{\link[=scr_test]{scr_test()}}: randomization test for the treatment effect.
\item \code{\link[=smart]{smart()}}: multi-arm extension.
}
}
\references{
Yang, H., Qin, Y., Li, Y., and Hu, F. (2024). Sequential
covariate-adjusted randomization via hierarchically minimizing
Mahalanobis distance and marginal imbalance. \emph{Biometrics}, 80(2),
ujae047. \doi{10.1093/biomtc/ujae047}
}
\author{
\strong{Maintainer}: Haoyu Yang \email{haoyuyang@hsph.harvard.edu}

Authors:
\itemize{
  \item Yichen Qin \email{qinyn@ucmail.uc.edu}
  \item Yang Li \email{yang.li@ruc.edu.cn}
  \item Feifang Hu
  \item Fanglu Chen \email{chenfanglu1031@qq.com} (original implementation in SCR 1.0.0)
}

}
\keyword{internal}
