% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot2d.R
\name{scr2d}
\alias{scr2d}
\title{S.C.R for 2d Data}
\usage{
scr2d(
  data,
  t = NULL,
  alpha = 0.05,
  h_hat = "AM",
  cent = TRUE,
  p = 4/3,
  grid_points = 201L,
  point_size = 3,
  point_color = "black",
  point_shape = 1,
  curve_size = 1.5,
  curve_color = "black",
  curve_linetype = "solid",
  scr_alpha = 1,
  scr_color = "#dfdfdf",
  scr_fill = "#dfdfdf",
  xlabel = "y1",
  ylabel = "y2",
  xlimit = NULL,
  ylimit = NULL,
  plot_title = "",
  aspect = TRUE
)
}
\arguments{
\item{data}{Matrix or data.frame with two columns.}

\item{t}{Vector whose length is the same as the number of rows in the data. This is not always necessary.}

\item{alpha}{Numeric. 1 - confidence coefficient.}

\item{h_hat}{"MAX", "AM", "GM" or "MIN". This parameter determines the choice of h in the Algorithms}

\item{cent}{Boolean.}

\item{aspect}{}
}
\description{
Construct S.C.R for 2d Data with Algorithms in the paper.
}
