% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr3_1.R
\name{scr3_1}
\alias{scr3_1}
\title{scr3_1}
\usage{
scr3_1(n = 100, lambda = 0.3, eta = 10)
}
\arguments{
\item{n}{sample size.}

\item{lambda}{Vector of rates.}

\item{eta}{Vector of shifts.}
}
\value{
plot
}
\description{
The solution to the exercise 3_1 for Rizzo's Book.

Write a function that will generate and return a randomsample of size n from the two-parameter
exponential distribution Exp(\eqn{\lambda,\eta}) for arbitraryn, \eqn{\lambda}, and \eqn{\eta}. (See Examples 2.3 and 2.6.)
Generate a large sample from Exp(\eqn{\lambda,\eta}) and compare the sample quantiles with the theoretical
quantiles.
}
\examples{
scr3_1(n=100,lambda=0.3,eta=10)

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
