% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr3_11.R
\name{scr3_11}
\alias{scr3_11}
\title{scr3_11}
\usage{
scr3_11(n = 1000, p = 0.75)
}
\arguments{
\item{n}{sample size}

\item{p}{probability---p1}
}
\value{
plot
}
\description{
The solution to the exercise 3_11 for Rizzo's Book.

 Generate a random sample of size 1000 from a normal location mixture. The components
 of the mixture have N(0,1)and N(3,1)distributions with mixing probabilities p1 and p2
 =1-p1. Graph the histogram of the sample with density superimposed, for p1 =0 .75.
 Repeat with different values for p1 and observe whether the empirical distribution
 of the mixture appears to be bimodal. Make a conjecture about the values of p1
 that produce bimodal mixtures.
}
\examples{
scr3_11(n=1000,p=0.75)

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
