% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr3_12.R
\name{scr3_12}
\alias{scr3_12}
\title{scr3_12}
\usage{
scr3_12(n = 1000, r = 4, beta = 2)
}
\arguments{
\item{n}{sample size}

\item{r}{an alternative way to specify the scale.}

\item{beta}{parmetor in gamma distribution.}
}
\value{
plot
}
\description{
The solution to the exercise 3_12 for Rizzo's Book.

Generate 1000 random observations from this mixture with r = 4 and \eqn{\beta} = 2.
}
\examples{
scr3_12(n=1000,r=4,beta=2)

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
