% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr3_14.R
\name{scr3_14}
\alias{scr3_14}
\title{scr3_14}
\usage{
scr3_14(n = 200)
}
\arguments{
\item{n}{sample size}
}
\value{
plot
}
\description{
The solution to the exercise 3_14 for Rizzo's Book.

Use the R pairs plot to graph an array of scatter plots for each pair of variables.
}
\examples{
scr3_14(n=200)

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
