% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr3_15.R
\name{scr3_15}
\alias{scr3_15}
\title{scr3_15}
\usage{
scr3_15(n = 100)
}
\arguments{
\item{n}{sample size}
}
\value{
list
}
\description{
The solution to the exercise 3_15 for Rizzo's Book.

Write a function that will standardize a multivariate normal sample for arbitrary n and d. That
is, transform the sample so that the sample mean vector is zero and sample covariance is the
identity matrix. To check your results, generate multivariate normal samples and print the
sample mean vector and covariance matrix before and after standardization.
}
\examples{
scr3_15(n=100)

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
