% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr3_16.R
\name{scr3_16}
\alias{scr3_16}
\title{scr3_16}
\usage{
scr3_16(n = 100)
}
\arguments{
\item{n}{sample size}
}
\value{
list
}
\description{
The solution to the exercise 3_16 for Rizzo's Book.

Efron and Tibshirani discuss the scor (bootstrap) test score data on 88 students who took
examinations in five subjects [84, Table 7.1], [188, Table 1.2.1]. Each row of the data frame
is a set of scores (xi1,...,x i5) for the ith student. Standardize the scores by type of exam.
That is, standardize the bivariate samples (X1,X2) (closed book) and the trivariate samples
(X3,X4,X5) (open book). Compute the covariance matrix of the transformed sample of test scores.
}
\examples{
scr3_16(n=200)

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
