% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr3_19.R
\name{scr3_19}
\alias{scr3_19}
\title{scr3_19}
\usage{
scr3_19(p = 0.5, s_min = 0, s_max = 20, s_start = 10)
}
\arguments{
\item{p}{the probability of one side of coin flips.}

\item{s_min}{the minimum dallors to stop the random walk.}

\item{s_max}{the maximum dallors to stop the random walk.}

\item{s_start}{the current dollors the player held.}
}
\value{
plot
}
\description{
The solution to the exercise 3_19 for Rizzo's Book.

Suppose that A and B each start with a stake of $10,and bet $1 on consecutive coin flips. The
game ends when either one of the players has all the money. Let Sn be the fortune of player A
at time n. Then {Sn,n \eqn{\ge} 0} is a symmetric random walk with absorbing barriers at 0 and 20.
Simulate a realization of the process {Sn,n \eqn{\ge} 0} and plot Sn vs the time index from time 0
until a barrier is reached.
}
\examples{
scr3_19(p=0.5,s_min=0,s_max=20,s_start=10)

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
