% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr3_2.R
\name{scr3_2}
\alias{scr3_2}
\title{scr3_2}
\usage{
scr3_2(n = 1000)
}
\arguments{
\item{n}{sample size}
}
\value{
plot
}
\description{
The solution to the exercise 3_2 for Rizzo's Book.

The standard Laplace distribution has density f(x)=1/2*e^-|x|, x \eqn{\in} R. Use the inverse transform
method to generate a random sample of size 1000 from this distribution.
Use one of the methods shown in this chapter to compare the generated sample to the target
distribution.
}
\examples{
scr3_2(n=1000)

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
