% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr3_20.R
\name{scr3_20}
\alias{scr3_20}
\title{scr3_20}
\usage{
scr3_20(n = 100)
}
\arguments{
\item{n}{the number of times the parameter sits changed.}
}
\value{
value
}
\description{
The solution to the exercise 3_20 for Rizzo's Book.

A compound Poisson process is a stochastic process {X(t),t \eqn{\ge} 0} that can be represented as the
random sum X(t), where {N(t),t\eqn{\ge} 0} is a Poisson process and Y1,Y2,... are iid
and independent of {N(t),t???0}.Write a program to simulate a compound Poisson(\eqn{\lambda})-Gamma
process (Y has a Gamma distribution). Estimate the mean and the variance of X(10) for several
choices of the parameters and compare with the theoretical values.
}
\examples{
scr3_20(n=200)

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
