% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr3_21.R
\name{scr3_21}
\alias{scr3_21}
\title{scr3_21}
\usage{
scr3_21(n = 100, in_min = 4, in_max = 5)
}
\arguments{
\item{n}{sample size}
}
\value{
list
}
\description{
The solution to the exercise 3_21 for Rizzo's Book.

A nonhomogeneous Poisson process has mean value function m(t)=t2 +2t, t \eqn{\ge} 0. Determine the
intensity function \eqn{\lambda} (t) of the process, and write a program to simulate the process on the
interval [4,5]. Compute the probability distributionof N(5)-N(4),and compare it to the
empirical estimate obtained by replicating the simulation.
}
\examples{
scr3_21(n=100)
}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
