% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr3_4.R
\name{scr3_4}
\alias{scr3_4}
\title{scr3_4}
\usage{
scr3_4(n = 1000, sigma = 1)
}
\arguments{
\item{n}{Sample size}

\item{sigma}{The parameter of Rayleigh.}
}
\value{
plot
}
\description{
The solution to the exercise 3_4 for Rizzo's Book.

Develop an algorithm to generate random samples from a Rayleigh(\eqn{\sigma}) distribution. Generate
Rayleigh(\eqn{\sigma}) samples for several choices \eqn{\sigma}f\eqn{\sigma} > 0 and check that the mode of the generated
samples is close to the theoretical mode &R (check the histogram).
}
\examples{
scr3_4(n=1000,sigma=1)

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
