% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr3_5.R
\name{scr3_5}
\alias{scr3_5}
\title{scr3_5}
\usage{
scr3_5(n = 1000)
}
\arguments{
\item{n}{sample size}
}
\value{
plot
}
\description{
The solution to the exercise 3_5 for Rizzo's Book.

Use the inverse transform method to generate a random sample of size 1000 from the
distribution of X. Construct a relative frequency table and compare the empirical
with the theoretical probabilities. Repeat using the R sample function.
}
\examples{
scr3_5(n=1000)

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
