% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr3_7.R
\name{scr3_7}
\alias{scr3_7}
\title{scr3_7}
\usage{
scr3_7(n = 1000, a = 3, b = 2)
}
\arguments{
\item{n}{sample size}

\item{a}{non-negative parameters of the Beta}

\item{b}{distribution}
}
\value{
plot
}
\description{
The solution to the exercise 3_7 for Rizzo's Book.

Write a function to generate a random sample of size n from the Beta(a,b)
distribution by the acceptance-rejection method. Generate a random sample of size 1000 from the
Beta(3,2) distribution. Graph the histogram of the sample with the theoretical
Beta(3,2) density superimposed.
}
\examples{
scr3_7(n=1000,a=3,b=2)

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
