% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr3_8.R
\name{scr3_8}
\alias{scr3_8}
\title{scr3_8}
\usage{
scr3_8(n = 1000, mu = 0, sigma = 1)
}
\arguments{
\item{n}{sample size}

\item{mu}{mean of the distribution on the log scale with default values of 0 and 1.}

\item{sigma}{standard deviation of the distribution on the log scale with default values of 0 and 1.}
}
\value{
plot
}
\description{
The solution to the exercise 3_8 for Rizzo's Book.

Write a function to generate random variates from a Lognormal(\eqn{\mu,\sigma}) distribution
using a transformation method, and generate a random sample of size 1000. Compare
the histogram with the lognormal density curve given by the dlnorm function in R.
}
\examples{
scr3_8(n=1000,mu=0,sigma=1)

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
