% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr3_9.R
\name{scr3_9}
\alias{scr3_9}
\title{scr3_9}
\usage{
scr3_9(n = 1000, a = -1, b = 1)
}
\arguments{
\item{n}{sample size}

\item{a}{lower limits of the distribution.}

\item{b}{upper limits of the distribution.}
}
\value{
plot
}
\description{
The solution to the exercise 3_9 for Rizzo's Book.

 Write a function to generate random variates from Fe, and construct the histogram 
 density estimate of a large simulated random sample.
}
\examples{
scr3_9(n=1000,a=-1,b=1)

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
