% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot3d.R
\name{scr3d}
\alias{scr3d}
\title{S.C.R for 3d Data}
\usage{
scr3d(
  data,
  t = NULL,
  div = 16,
  alpha = 0.05,
  h_hat = "AM",
  cent = TRUE,
  p = 2,
  grid_points = 201L,
  point_size = 1,
  point_color = "black",
  curve_size = 4,
  curve_color = "black",
  scr_alpha = 0.4,
  scr_color = "#afafaf",
  xlabel = "y1",
  ylabel = "y2",
  zlabel = "y3"
)
}
\arguments{
\item{data}{Matrix or data.frame with three columns.}

\item{t}{Vector whose length is the same as the number of rows in the data. This is not always necessary.}

\item{zlabel}{}
}
\description{
Construct S.C.R for 3d Data with Algorithms in the paper.
}
