% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr4_1.R
\name{scr4_1}
\alias{scr4_1}
\title{scr4_1}
\usage{
scr4_1(n = 200)
}
\arguments{
\item{n}{sample size}
}
\value{
plot
}
\description{
The solution to the exercise 4_1 for Rizzo's Book.

Use the R pairs plot to graph an array of scatter plots for each pair of variables.
}
\examples{
scr4_1(n=200)

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
