% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr4_11.R
\name{scr4_11}
\alias{scr4_11}
\title{scr4_11}
\usage{
scr4_11()
}
\value{
plot
}
\description{
The solution to the exercise 4_11 for Rizzo's Book.

Refer to the full leafshape (DAAG) data set. Display a segment style stars plot for leaf
measurements at latitude 42 (Tasmania). Repeat using the logarithms of the measurements.
}
\examples{
scr4_11()

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
