% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr4_2.R
\name{scr4_2}
\alias{scr4_2}
\title{scr4_2}
\usage{
scr4_2(n = 200)
}
\arguments{
\item{n}{sample size}
}
\value{
plot
}
\description{
The solution to the exercise 4_2 for Rizzo's Book.

Add a fitted smoothcurve to each of the scatterplotsin Figure4.1 of Example 4.1.
}
\examples{
scr4_2(n=200)

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
