% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr4_3.R
\name{scr4_3}
\alias{scr4_3}
\title{scr4_3}
\usage{
scr4_3(n = 200, p = 0.75)
}
\arguments{
\item{n}{sample size}

\item{p}{the probability of the first kind of normal distribution.}
}
\value{
plot
}
\description{
The solution to the exercise 4_3 for Rizzo's Book.

Generate a bivariate random sample from the joint distribution of (X,Y)
and construct a contour plot.
}
\examples{
scr4_3(n=200,p=0.75)

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
