% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr4_4.R
\name{scr4_4}
\alias{scr4_4}
\title{scr4_4}
\usage{
scr4_4(n = 200, p = 0.75)
}
\arguments{
\item{n}{sample size}

\item{p}{the probability of the first kind of normal distribution.}
}
\value{
plot
}
\description{
The solution to the exercise 4_4 for Rizzo's Book.

Construct a filled contour plot of the bivariate mixture in Exercise 4.3.
}
\examples{
scr4_4(n=200,p=0.75)

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
