% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr4_7.R
\name{scr4_7}
\alias{scr4_7}
\title{scr4_7}
\usage{
scr4_7(n = 200, a = 1)
}
\arguments{
\item{n}{the number of observations}

\item{a}{the initial plot of the data when a=1, the adjusted data when a=0.}
}
\value{
plot
}
\description{
The solution to the exercise 4_7 for Rizzo's Book.

Create a parallel coordinates plot of the crabs (MASS)[278] data using all 200
observations. Compare the plots before and after adjusting the measurements by the
size of the crab. Interpret the resulting plots.
}
\examples{
scr4_7(n=200)

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
