% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr4_8.R
\name{scr4_8}
\alias{scr4_8}
\title{scr4_8}
\usage{
scr4_8()
}
\value{
plot
}
\description{
The solution to the exercise 4_8 for Rizzo's Book.

Create a plot of Andrews curves for the leafshape17 (DAAG) [185] data, using the logarithms of
measurements (logwid, logpet, loglen). Set line type to identify leaf architecture as in
Example 4.9. Compare with the plot in Figure 4.8.
}
\examples{
scr4_8()
}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
