% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr4_9.R
\name{scr4_9}
\alias{scr4_9}
\title{scr4_9}
\usage{
scr4_9()
}
\value{
plot
}
\description{
The solution to the exercise 4_9 for Rizzo's Book.

Refer to the full leafshape (DAAG) data set. Produce Andrews curves for each of the six
locations. Split the screen into six plotting areas, and display all six plots on one screen.
Set line type or color to identify leaf architecture.
}
\examples{
scr4_9()

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
