% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr5_1.R
\name{scr5_1}
\alias{scr5_1}
\title{scr5_1}
\usage{
scr5_1(m = 1000)
}
\arguments{
\item{m}{a number indicating iteration numbers. Defaults to 1000.}
}
\value{
ans a data.frame containing the the Monte Carlo estimate and the exact value of the
integral.
}
\description{
The solution to Rizzo's book exercise 5.1.

Compute a Monte Carlo estimate of \eqn{int_{0}^{\pi/3} sint dt} and compare
your estimate with the exact value of the integral.
}
\examples{
scr5_1(10000)

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
