% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr5_10.R
\name{scr5_10}
\alias{scr5_10}
\title{scr5_10}
\usage{
scr5_10(m = 1000)
}
\arguments{
\item{m}{a number indicating the iteration numbers. Defaults to 1000.}
}
\value{
data.frame containing the Monte Carlo estimate anf the Monte Carlo intergration with
antithetic variables estimate of the integral, and the variance reduction percentage.
}
\description{
The solution to the exercise 5_10 for Rizzo's Book.

Use Monte Carlo integration with antithetic variables to estimate
\eqn{\int_0^1 \frac{e^{-x}}{1 + x^2}}
and ﬁnd the approximate reduction in variance as a percentage of the variance
without variance reduction.
}
\examples{
scr5_10()


}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
