% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr5_14.R
\name{scr5_14}
\alias{scr5_14}
\title{scr5_14}
\usage{
scr5_14(m = 1000)
}
\arguments{
\item{m}{a number indicating the number of iterations. Defaults to 1000.}
}
\value{
data.frame containing the Monte Carlo estimate of theta hat by
importantce sampling, and the estimation variance.
}
\description{
The solution to Rizzo's book exercise 5.14.

Obtain a Monte Carlo estimate of \eqn{\int_1^{\infinity}\frac{x^2}{\sqrt(2\pi)}e^{-x^2/2}dx}
by importance sampling.
}
\examples{
scr5_14(m = 5000)



}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
