% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr5_15.R
\name{scr5_15}
\alias{scr5_15}
\title{scr5_15}
\usage{
scr5_15(m = 1000)
}
\arguments{
\item{m}{a number indicating the iteration numbers. Defaults to 1000.}
}
\value{
data.frame containing the estimates of theta hat and estimation variances
corresponding to example 5.10 and 5.13 respectively.
}
\description{
The solution of Rizzo's book exercise 5.15.

Obtain the stratiﬁed importance sampling estimate in Example 5.13
and compare it with the result of Example 5.10.
}
\examples{
scr5_15(m = 5000)



}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
