% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr5_2.R
\name{scr5_2}
\alias{scr5_2}
\title{scr5_2}
\usage{
scr5_2(x = 2, m = 1000)
}
\arguments{
\item{x}{a list of number(s), corresponding to the x in the exercise description.
Defaults to 2.}

\item{m}{a number indicating the iteration numbers. Defaults to 1000.}
}
\value{
ans a data.frame consists with the Monte Carlo estimates of Φ(x) and the estimates
with normal cdf function `pnorm`.
}
\description{
The solution to Rizzo's Book exercise 5.2.

Compute a Monte Carlo estimate of the standard normal cdf,
by generating from the Uniform(0,x) distribution.
Compare your estimates with the normal cdf function pnorm.
Compute an estimate of the variance of your Monte Carlo estimate of Φ(2),
and a 95\% conﬁdence interval for Φ(2).
}
\examples{
scr5_2(1:5, 1000)

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
