% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr5_3.R
\name{scr5_3}
\alias{scr5_3}
\title{scr5_3}
\usage{
scr5_3(m = 1000)
}
\arguments{
\item{m}{a number indicating the iteration numbers. Defaults to 1000.}
}
\value{
a data.frame containing the true value of theta, the two Monte Carlo estimates
(theta.hat and theta.star), and their corresponding variance estimates.
}
\description{
The solution to Rizzo's book exercise 5.3.

Compute a Monte Carlo estimate θ of \eqn{\thata = \int_0^{0.5}e^{-x}dx} by sampling from
Uniform(0, 0.5), and estimate the variance of \eqn{\hat{θ}}. Find another Monte Carlo estimator
{θ^∗} by sampling from the exponential distribution.
Which of the variances (of \eqn{\hat{θ}} and {θ^∗}) is smaller, and why?
}
\examples{
scr5_3(10000)


}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
