% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr5_5.R
\name{scr5_5}
\alias{scr5_5}
\title{scr5_5}
\usage{
scr5_5(m = 10000, x = seq(0.1, 2.5, length = 10))
}
\arguments{
\item{m}{a number indicating the iteration numbers. Defaults to 1000.}

\item{x}{a list of number(s) containing the interested x value corresponding to the
problem.}
}
\value{
data.frame contains the true value of the standard normal cdf, the Monte Carlo
estimate, the Hit or Miss estimate, the variances corresponding to these two methods,
and their efficnency.
}
\description{
The solution to Rizzo's book exercise 5_5.

Compute (empirically) the eﬃciency of the sample mean Monte Carlo method of
estimation of the deﬁnite integral in Example 5.3 (Use the Monte Carlo approach
to estimate the standard normal cdf) relative to the “hit or miss” method.
}
\examples{
scr5_5()


}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
