% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr5_6.R
\name{scr5_6}
\alias{scr5_6}
\title{scr5_6}
\usage{
scr5_6()
}
\value{
data.frame contains the covariance and variance corresponding to the problem,
and the variance reduction percent.
}
\description{
The solution to Rizzo's book exercise 5_6.

In Example 5.7 the control variate approach was illustrated for an Monte Carlo integration.
Now consider the antithetic variate approach. Compute Cov(e^U , e^{1−U}) and
Var(e^U + e^{1−U} ), where U ∼ Uniform(0,1).
What is the percent reduction in variance of θ that can be achieved using
antithetic variates (compared with simple MC)?
}
\examples{
scr5_6()


}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
