% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr5_7.R
\name{scr5_7}
\alias{scr5_7}
\title{scr5_7}
\usage{
scr5_7(m = 1000)
}
\arguments{
\item{m}{a number indicating the iteration numbers. Defaults to 1000.}
}
\value{
data.frame containing the empirical estimate of the variance reduction using
the antithetic variate, and the theoretical value from Exercise 5.6.
}
\description{
The solution to the Rizzo's book exercise 5_7.

Refer to Exercise 5.6. Use a Monte Carlo simulation to estimate θ by the
antithetic variate approach and by the simple Monte Carlo method.
Compute an empirical estimate of the percent reduction in variance using the
antithetic variate. Compare the result with the theoretical value from Exercise 5.6.
}
\examples{
scr5_7()


}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
