% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr6_1.R
\name{scr6_1}
\alias{scr6_1}
\title{scr6_1}
\usage{
scr6_1(n = 20, k = n/2 - 1, m = 1000)
}
\arguments{
\item{n}{random samples size generated from a standard Cauchy distribution. Defaults to 20}

\item{k}{level k trimmed means}

\item{m}{number of response vectors to simulate. Defaults to 1000}
}
\value{
matrix
}
\description{
The solution to the exercise 6_1 for Rizzo's Book.

Estimate the MSE of the level k trimmed means for random samples of size 20 generated from a standard Cauchy distribution. (The target parameter θ is the center or median; the expected value does not exist.) Summarize the estimates of MSE in a table for k = 1, 2, . . . , 9.
}
\examples{
scr6_1(n=20,k=n/2-1,m=1000)

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
