% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr6_10.R
\name{scr6_10}
\alias{scr6_10}
\title{scr6_10}
\usage{
scr6_10(n = 50, m = 1000)
}
\arguments{
\item{n}{random samples size. Defaults to 50}

\item{m}{number of response vectors to simulate. Defaults to 1000}
}
\description{
The solution to the exercise 6_10 for Rizzo's Book.

Construct an approximate 95% confidence interval for the Gini ratio γ = E[G] if X is lognormal with unknown parameters. Assess the coverage rate of the estimation procedure with a Monte Carlo experiment.
}
\examples{
scr6_10(n=50,m=1000)

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
