% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr6_2.R
\name{scr6_2}
\alias{scr6_2}
\title{scr6_2}
\usage{
scr6_2(n = 20, m = 1000, mu0 = 500)
}
\arguments{
\item{n}{random samples size. Defaults to 20}

\item{m}{number of response vectors to simulate. Defaults to 1000}

\item{mu0}{null hypothesis. Defaults to 500}
}
\value{
plot
}
\description{
The solution to the exercise 6_2 for Rizzo's Book.

Plot the empirical power curve for the t-test in Example 6.9, changing the alternative hypothesis to H1 : μ ≠ 500, and keeping the significance level α = 0.05.
}
\examples{
scr6_2(n=20,m=1000, mu0=500)

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
