% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr6_3.R
\name{scr6_3}
\alias{scr6_3}
\title{scr6_3}
\usage{
scr6_3()
}
\value{
plot
}
\description{
The solution to the exercise 6_3 for Rizzo's Book.

Plot the power curves for the t-test in Example 6.9 for sample sizes 10, 20, 30, 40, and 50, but omit the standard error bars. Plot the curves on the same graph, each in a different color or different line type, and include a legend. Comment on the relation between power and sample size.
}
\examples{
scr6_3()

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
