% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr6_4.R
\name{scr6_4}
\alias{scr6_4}
\title{scr6_4}
\usage{
scr6_4(n = 1000, mu = 0, sigma = 1, m = 1000, alpha = 0.05)
}
\arguments{
\item{n}{random samples size generated from a lognormal distribution. Defaults to 1000}

\item{mu}{sample mean. Defaults to 0.}

\item{sigma}{ssample standard deviation. Defaults to 1.}

\item{m}{number of response vectors to simulate. Defaults to 1000}

\item{alpha}{significance level. Defaults to 0.05}
}
\description{
The solution to the exercise 6_4 for Rizzo's Book.

Suppose that X1, . . . , Xn are a random sample from a from a lognormal distribution with unknown parameters. Construct a 95% confidence interval for the parameter μ. Use aMonte Carlomethod to obtain an empirical estimate of the confidence level.
}
\examples{
scr6_4(n=1000,mu=500,m=1000,alpha=0.05)
}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
