% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr6_6.R
\name{scr6_6}
\alias{scr6_6}
\title{scr6_6}
\usage{
scr6_6(n = 20, m = 1000)
}
\arguments{
\item{n}{random samples size. Defaults to 20}

\item{m}{number of response vectors to simulate. Defaults to 1000}
}
\value{
matrix
}
\description{
The solution to the exercise 6_1 for Rizzo's Book.

Estimate the 0.025, 0.05, 0.95, and 0.975 quantiles of the skewness √b1 under normality by a Monte Carlo experiment. Compute the standard error of the estimates from (2.14) using the normal approximation for the density (with exact variance formula). Compare the estimated quantiles with the quantiles of the large sample approximation √b1 ≈ N(0, 6/n).
}
\examples{
scr6_6(n=20,m=1000)

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
