% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr6_7.R
\name{scr6_7}
\alias{scr6_7}
\title{scr6_7}
\usage{
scr6_7(n = 20, m = 1000, alpha = 0.05)
}
\arguments{
\item{n}{random samples size. Defaults to 20}

\item{m}{number of response vectors to simulate. Defaults to 1000}

\item{alpha}{significance level. Defaults to 0.05}
}
\description{
The solution to the exercise 6_7 for Rizzo's Book.

Estimate the power of the skewness test of normality against symmetric Beta(α, α) distributions and comment on the results.
}
\examples{
scr6_7(n=20,m=1000,alpha=0.05)

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
