% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr6_8.R
\name{scr6_8}
\alias{scr6_8}
\title{scr6_8}
\usage{
scr6_8()
}
\value{
matrix
}
\description{
The solution to the exercise 6_8 for Rizzo's Book.

Refer to Example 6.16. Repeat the simulation, but also compute the F test of equal variance, at significance level α = 0.055. Compare the power of the Count Five test and F test for small, medium, and large sample sizes. (Recall that the F test is not applicable for non-normal distributions.)
}
\examples{
scr6_8()

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
