% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr6_9.R
\name{scr6_9}
\alias{scr6_9}
\title{scr6_9}
\usage{
scr6_9(n = 100, m = 10000)
}
\arguments{
\item{n}{random samples size. Defaults to 100}

\item{m}{number of response vectors to simulate. Defaults to 10000}
}
\description{
The solution to the exercise 6_9 for Rizzo's Book.

Let X be a non-negative random variable with μ = E[X] < ∞. For a random sample x1, . . . , xn from the distribution of X. The Gini ratio is applied in economics to measure inequality in income distribution (see e.g. [163]). If the mean is unknown, let ˆG be the statistic G with μ replaced by ¯x. Estimate by simulation the mean, median and deciles of ˆG if X is standard lognormal. Repeat the procedure for the uniform distribution and Bernoulli(0.1). Also construct density histograms of the replicates in each case.
}
\examples{
scr6_9(n=100,m=10000)

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
