% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr7_1.R
\name{scr7_1}
\alias{scr7_1}
\title{scr7_1}
\usage{
scr7_1()
}
\description{
The solution to the exercise 7_1 for Rizzo's Book.

Compute a jackknife estimate of the bias and the standard error of the correlation statistic in Example 7.2.
}
\examples{
scr7_1()
}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
