% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr7_2.R
\name{scr7_2}
\alias{scr7_2}
\title{scr7_2}
\usage{
scr7_2(B = 200)
}
\arguments{
\item{B}{number of replicates.  Defaults to 200}
}
\description{
The solution to the exercise 7_2 for Rizzo's Book.

Refer to the law data (bootstrap). Use the jackknife-after-bootstrapmethod to estimate the standard error of the bootstrap estimate of se(R).
}
\examples{
scr7_2(B=200)
}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
