% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr7_3.R
\name{scr7_3}
\alias{scr7_3}
\title{scr7_3}
\usage{
scr7_3(B = 200, alpha = 0.05)
}
\arguments{
\item{B}{number of replicates.  Defaults to 200}

\item{alpha}{significance level}
}
\description{
The solution to the exercise 7_3 for Rizzo's Book.

Obtain a bootstrap t confidence interval estimate for the correlation statistic in Example 7.2 (law data in bootstrap).
}
\examples{
scr7_3(B=200,alpha=0.05)
}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
