% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr7_4.R
\name{scr7_4}
\alias{scr7_4}
\title{scr7_4}
\usage{
scr7_4(B = 200)
}
\arguments{
\item{B}{number of replicates.  Defaults to 200}
}
\description{
The solution to the exercise 7_4 for Rizzo's Book.

Refer to the air-conditioning data set aircondit provided in the boot package. The 12 observations are the times in hours between failures of airconditioning equipment [63, Example 1.1]: 3, 5, 7, 18, 43, 85, 91, 98, 100, 130, 230, 487. Assume that the times between failures follow an exponential model Exp(λ). Obtain the MLE of the hazard rate λ and use bootstrap to estimate the bias and standard error of the estimate.
}
\examples{
scr7_4(B=200)
}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
