% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr7_5.R
\name{scr7_5}
\alias{scr7_5}
\title{scr7_5}
\usage{
scr7_5(B = 200)
}
\arguments{
\item{B}{number of replicates.  Defaults to 200}
}
\description{
The solution to the exercise 7_5 for Rizzo's Book.

Refer to Exercise 7.4. Compute 95% bootstrap confidence intervals for the mean time between failures 1/λ by the standard normal, basic, percentile, and BCa methods. Compare the intervals and explain why they may differ.
}
\examples{
scr7_5(B=200)
}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
