% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr7_7.R
\name{scr7_7}
\alias{scr7_7}
\title{scr7_7}
\usage{
scr7_7(B = 200)
}
\arguments{
\item{B}{number of replicates.  Defaults to 200}
}
\description{
The solution to the exercise 7_7 for Rizzo's Book.

Refer to Exercise 7.6. Efron and Tibshirani discuss the following example [84, Ch. 7]. The five-dimensional scores data have a 5 × 5 covariance matrix Σ, with positive eigenvalues λ1 > · · · > λ5. Let ˆλ1 > · · · > ˆλ5 be the eigenvalues of ˆΣ, where ˆΣ is the MLE of Σ. Compute the sample estimate of θ. Use bootstrap to estimate the bias and standard error of ˆθ.
}
\examples{
scr7_7(B=200)
}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
