% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr7_8.R
\name{scr7_8}
\alias{scr7_8}
\title{scr7_8}
\usage{
scr7_8(B = 200)
}
\arguments{
\item{B}{number of replicates.  Defaults to 200}
}
\description{
The solution to the exercise 7_8 for Rizzo's Book.

Refer to Exercise 7.7. Obtain the jackknife estimates of bias and standard error of ˆθ.
}
\examples{
scr7_8(B=200)
}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
