% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr7_9.R
\name{scr7_9}
\alias{scr7_9}
\title{scr7_9}
\usage{
scr7_9(B = 200)
}
\arguments{
\item{B}{number of replicates.  Defaults to 200}
}
\description{
The solution to the exercise 7_9 for Rizzo's Book.

Refer to Exercise 7.7. Compute 95% percentile and BCa confidence intervals for ˆθ.
}
\examples{
scr7_9(B=200)
}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
