% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr8_1.R
\name{scr8_1}
\alias{scr8_1}
\title{scr8_1}
\usage{
scr8_1(rep = 1000)
}
\arguments{
\item{rep}{number of replicates.  Defaults to 1000}
}
\description{
The solution to the exercise 8_1 for Rizzo's Book.

Implement the two-sample Cram´er-von Mises test for equal distributions as a permutation test. Apply the test to the data in Examples 8.1 and 8.2.
}
\examples{
scr8_1(rep=1000)
}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
