% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr8_2.R
\name{scr8_2}
\alias{scr8_2}
\title{scr8_2}
\usage{
scr8_2(rep = 1000)
}
\arguments{
\item{rep}{number of replicates.  Defaults to 1000}
}
\description{
The solution to the exercise 8_2 for Rizzo's Book.

Implement the bivariate Spearman rank correlation test for independence [255] as a permutation test. The Spearman rank correlation test statistic can be obtained from function cor with method = "spearman". Compare the achieved significance level of the permutation test with the p-value reported by cor.test on the same samples.
}
\examples{
scr8_2(rep=1000)
}
\references{
Maria L. Rizzo. (2016). Statistical computing with rep.cor.
}
