% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr9_1.R
\name{scr9_1}
\alias{scr9_1}
\title{scr9_1}
\usage{
scr9_1(m = 10000, sigma = 2)
}
\arguments{
\item{m}{a number indicating the iteration numbers. Defaults to 10000.}

\item{sigma}{a number indicating the sigma parameter related to Rayleigh distribution.}
}
\value{
plot showing part of a chain generated by a Metropolis-Hastings sampler
for the target Rayleigh distribution with σ = 2.
}
\description{
The solution to the exercise 9_1 for Rizzo's Book.

Repeat Example 9.1 for the target distribution Rayleigh(σ = 2).
Compare the performance of the Metropolis-Hastings sampler for Example 9.1
and this problem. In particular, what diﬀerences are obvious from the
plot corresponding to Figure 9.1?
}
\examples{
scr9_1()



}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
