% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr9_10.R
\name{scr9_10}
\alias{scr9_10}
\title{scr9_10}
\usage{
scr9_10(n = 2000, b = 500, x0 = c(2, 3, 4, 5), sigma = 4)
}
\arguments{
\item{n}{a number indicating the length of the chain. Defaults to be 15000.}

\item{b}{a number indicating the burn-in time. Defaults to be 1000.}

\item{x0}{a list of number(s) indicating the overdispersed initial values.
Defaults to [2, 3, 4, 5]}
}
\value{
r.hat a list of numbers indicating the r hat

convergence_num a number indicating the length of the chain until it reaches the
convergence.

plot showing the changing in r hat.
}
\description{
The solution to the exercise 9_10 for Rizzo's Book.

Refer to Example 9.1. Use the Gelman-Rubin method to monitor convergence of
the chain, and run the chain until the chain has converged approximately to
the target distribution according to R < 1.2. (See Exercise 9.9.) Also use
the coda [212] package to check for convergence of the chain by the
Gelman-Rubin method.
}
\examples{
scr9_10()



}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
