% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr9_11.R
\name{scr9_11}
\alias{scr9_11}
\title{scr9_11}
\usage{
scr9_11(n = 10000, burn = 50, w = c(0.1, 0.3, 0.5, 0.8))
}
\arguments{
\item{n}{a number indicating the length of the chain. Defaults to be 15000.}

\item{burn}{a number indicating the burn-in time. Defaults to be 1000.}

\item{w}{a list of numbers indicating the parameter of proposal distribution.
Defaluts to [0.1, 0.3, 0.5, 0.8]}
}
\value{
r.hat a list of numbers containing all the r hat.

convergence_num a number showing the length of the chain when it reaches the convergence.

fig a plot showing the r hat.
}
\description{
The solution to the exercise 9_11 for Rizzo's Book.

Refer to Example 9.5. Use the Gelman-Rubin method to monitor convergence of the
chain, and run the chain until the chain has converged approximately to the
target distribution according to R < 1.2.
}
\examples{
scr9_11()



}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
