% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr9_12.R
\name{scr9_12}
\alias{scr9_12}
\title{scr9_12}
\usage{
scr9_12(n = 15000, burn = 50, x0 = c(0.1, 0.2, 0.3, 0.4))
}
\arguments{
\item{n}{lenght of the chain. Defaults to be 15000.}

\item{burn}{burn-in time. Defaults to be 1000.=}

\item{x0}{overdispersed initial values. Defaults to be [0.1, 0.2, 0.3, 0.4]}
}
\value{
r.hat a list of numbers indicating the r hat

convergence_num a number indicating the length of the chain until it reaches the
convergence.

plot showing the changing in r hat.
}
\description{
The solution to the exercise 9_12 for Rizzo's Book.

Refer to Example 9.6. Use the Gelman-Rubin method to monitor convergence of the
chain, and run the chain until the chain has converged approximately to the
target distribution according to R < 1.2.
}
\examples{
scr9_12()



}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
