% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr9_2.R
\name{scr9_2}
\alias{scr9_2}
\title{scr9_2}
\usage{
scr9_2(m = 10000, sigma = 4)
}
\arguments{
\item{m}{a number indicating the iteration numbers. Defaults to 10000.}

\item{sigma}{a number indicating to the sigma parameter related to Rayleigh distribution.
Defaults to 4.}
}
\value{
plot showing the chain generated with Metropolis-Hasting sampler
for the target Rayleigh's distribution with sigma (defaluts to 4).
}
\description{
The solution to the exercise 9_2 for Rizzo's Book.

Repeat Example 9.1 using the proposal distribution Y
(shape parameter Xt and rate parameter 1).
}
\examples{
scr9_2()



}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
