% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr9_3.R
\name{scr9_3}
\alias{scr9_3}
\title{scr9_3}
\usage{
scr9_3(m = 10000)
}
\arguments{
\item{m}{a number indicating the iteration numbers. Defaults to 10000.}
}
\value{
data.frame containing the deciles of the generated observations and the
deciles of the standard Cauchy distribution.
}
\description{
The solution to the exercise 9_3 for Rizzo's Book.

Use the Metropolis-Hastings sampler to generate random variables
from a standard Cauchy distribution. Discard the ﬁrst 1000 of the chain,
and compare the deciles of the generated observations with the deciles of
the standard Cauchy distribution (see qcauchy or qt with df=1).
The standard Cauchy has the Cauchy(θ = 1, η = 0) density.
(Note that the standard Cauchy density is equal to the Student t
density with one degree of freedom.)
}
\examples{
scr9_3()



}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
