% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr9_5.R
\name{scr9_5}
\alias{scr9_5}
\title{scr9_5}
\usage{
scr9_5(m = 5000, b = 0.2, burn = 1000, w = 0.25, sed = 135)
}
\arguments{
\item{m}{a number indicating the length of the chain. Defaults to be 5000.}

\item{b}{a number indicating the actual value of data. Defaults to be 0.2.}

\item{burn}{a number indicating the burn-in time. Defaults to be 1000.}

\item{w}{a number indicating the width of the uniform support set. Defaults to be 0.25.}

\item{sed}{a number indicating the random seed number. Defaults to be 135.}
}
\value{
sample_mean a number indicating the sample mean of the generated chain.

plot showing the generated chain.
}
\description{
The solution to the exercise 9_5 for Rizzo's Book.

What eﬀect, if any, does the width w have on the mixing of the chain
in Example 9.5? Repeat the simulation keeping the random number seed ﬁxed,
trying diﬀerent proposal distributions based on the random increments from
Uniform(−w, w), varying w.
}
\examples{
scr9_5()



}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
