% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr9_6.R
\name{scr9_6}
\alias{scr9_6}
\title{scr9_6}
\usage{
scr9_6(m = 10000, burn = 1000)
}
\arguments{
\item{m}{a number indicating the length of the chain. Defaults to be 10000.}

\item{burn}{a number indicating the burn-in time. Defaults to be 1000.}
}
\value{
`estimate theta hat` a number indicating the estimate theta.

method a character string indicating which method is used.

data.name a character string giving the names of the data

x a list of number containing the generated normal chain Xt.

reject_rate a number indicating the reject rate in this method.

plot showing the generated chain.
}
\description{
The solution to the exercise 9_6 for Rizzo's Book.

Rao [220, Sec. 5g] presented an example on genetic linkage of 197 animals in
four categories. The group sizes are (125, 18, 20, 34). Assume that the probabilities of the
corresponding multinomial distribution are ((2 + θ)/4, (1 - θ)/4, (1 - θ)/4, θ/4).
Estimate the posterior distribution of θ given the observed sample,
using one of the methods in this chapter.
}
\details{
The results are generated with the Metropolis-Hasting Method.
More details about this exercise could be found in the expanded function
for this Chapter by referring `MCMCsim9`
}
\examples{
scr9_6()

}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
