% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr9_7.R
\name{scr9_7}
\alias{scr9_7}
\title{scr9_7}
\usage{
scr9_7(N = 5000, burn = 1000)
}
\arguments{
\item{N}{a number indicating the length of the chain. Defaults to be 5000.}

\item{burn}{a number indicating the burn-in time. Defaults to be 1000.}
}
\value{
x a list of numbers containing the generated normal chain Xt.

y a list of numbers containing the generated normal chain Yt.

lm_reg lthe inear regression result of the model Y = β0 + β1 X.

fig a scatter plot of the generated sample.
}
\description{
The solution to the exercise 9_7 for Rizzo's Book.

Implement a Gibbs sampler to generate a bivariate normal chain
(Xt, Yt ) with zero means, unit standard deviations, and correlation 0.9.
Plot the generated sample after discarding a suitable burn-in sample.
Fit a simple linear regression model Y = β0 + β1 X to the sample and
check the residuals of the model for normality and constant variance.
}
\examples{
scr9_7()
}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
