% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr9_8.R
\name{scr9_8}
\alias{scr9_8}
\title{scr9_8}
\usage{
scr9_8(N = 5000, burn = 1000)
}
\arguments{
\item{N}{a number indicating the length of the chain. Defaults to be 5000.}

\item{burn}{a number indicating the burn-in time. Defaults to be 1000.}
}
\value{
plot showing the scatter plot of the chain generated from Gibbs sampler.
}
\description{
The solution to the exercise 9_8 for Rizzo's Book.

Use the Gibbs sampler to generate a chain with target joint density f(x, y).
(f(x, y) can be found in Rizzo's book page 278).
}
\examples{
scr9_8()
}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
