% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr9_9.R
\name{scr9_9}
\alias{scr9_9}
\title{scr9_9}
\usage{
scr9_9(n = 15000, b = 1000, sigma = 0.2, x0 = c(-10, -5, 5, 10))
}
\arguments{
\item{n}{a number indicating the length of the chain. Defaults to be 15000.}

\item{b}{a number indicating the burn-in time. Defaults to be 1000.}

\item{sigma}{a number inficating the parameter of the proposal distribution. Defaluts to be 0.2.}

\item{x0}{a list of numbers indicating the overdispersed initial values.
Defaults to be [-10, -5, 5, 10].}
}
\value{
R.hat a number indicating the estimated R.
}
\description{
The solution to the exercise 9_9 for Rizzo's Book.

Modify the Gelman-Rubin convergence monitoring given in Example 9.8
so that only the ﬁnal value of R.hat is computed, and repeat the example,
omitting the graphs.
}
\examples{
scr9_9()



}
\references{
Maria L. Rizzo. (2016). Statistical computing with r.
}
