% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scType.R
\name{scType_pipeline}
\alias{scType_pipeline}
\title{scType}
\usage{
scType_pipeline(
  path_to_db_file = NULL,
  cell_types,
  srtObject,
  assay = "RNA",
  slot = "scale.data",
  scaled = scaled,
  resolution = "integrated_snn_res.1.2"
)
}
\arguments{
\item{path_to_db_file}{list of markers genes. If NULL the internal object will be used.}

\item{cell_types}{Tissue type. Used to select the markers.}

\item{srtObject}{input data, either a matrix or a seurat object}

\item{assay}{Assay from seurat}

\item{slot}{Slot from seurat}

\item{scaled}{TRUE if input data is already scaled}

\item{resolution}{clustering resolution to pass to scType}
}
\description{
Fully-automated and ultra-fast cell-type identification using specific marker 
combinations from single-cell transcriptomic data.
The functions were modified and adapted from the original repository
(https://github.com/IanevskiAleksandr/sc-type/tree/master).
}
\details{
* Written by Aleksandr Ianevski <aleksandr.ianevski@helsinki.fi>, June 2021
* Adapted by Francesco Carbone <francesco.carbone@institutiamgine.org>, July 2023
}
\examples{
\dontrun{scType_pipeline()}

}
\keyword{annotation}
