% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{setup_sc_project}
\alias{setup_sc_project}
\title{Setup project directory structure}
\usage{
setup_sc_project(project_location = getwd(), project_name = "new_project")
}
\arguments{
\item{project_location}{Directory in which the project will be located.  Default is "getwd()"}

\item{project_name}{Name of the project}
}
\value{
no return value.  Creates system folders.
}
\description{
Create reproducible project directory organization when initiating a new analysis.
}
\examples{
\dontrun{
# If using built-in directory structure.
setup_sc_project("/home/project","new_project")
}

}
\concept{organization_util}
