% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{stemnet_plot}
\alias{stemnet_plot}
\title{Plot a STEMNET object by Circular A Posteriori Projection}
\usage{
stemnet_plot(x, arrangement = NULL, top = 1, color = "direction")
}
\arguments{
\item{x}{An object of class \code{\link{stemnet}}}

\item{arrangement}{The order of developmental endpoints on the CAP plot. If \code{NULL} (default), the optimal arrangement is determined using a travelling salesman algorithm, see below.}

\item{top}{If \code{arrangement} is \code{NULL}, show the top n solutions for arranging the developmental endpoints. Defaults to 1, i.e. show only the best solution.}

\item{color}{Specify the variable color-coded on the plot. The following options exist:
\itemize{
\item One of "direction" or "degree": Highlight the degree or direction of lineage priming.
\item A numeric vector of length equal the number of cells included in the STEMNET plot
\item A boolean vector to impose a density plot on the STEMNET plot
}}
}
\value{
Invisibly returns a list of two elements: First, the data underlying the plot; Second, the arrangement of endpoints on the cirlce.
}
\description{
Plot a STEMNET object by Circular A Posteriori Projection
}
\details{
The posterior probability matrix computed by STEMNET places each cell into a n-simplex with as many corners as there a developmental endpoints. To reduce this high-dimensional space to 2 dimensions for plotting, the endpoints are arranged on the edge of a circle and all cells are placed in between.

In brief, each endpoint is assigned with an angle \eqn{\alpha_k}. The class probabilties \eqn{p_{ik}} are then transformed to cartesian coordinates by \deqn{x_i = \sum_k p_{ik} \cos (\alpha_k) } and  \deqn{y_i = \sum_k p_{ik} \sin (\alpha_k) }

To find the optimal arrangement of the developmental endpoints on the circle, lineages with common precursor stages are placed next to each other. The proximity between lineages l and k computed by \deqn{D_{kl} = \sum_i p_{il} * p_{ik}}

The arrangement with the highest proximity is chosen.
}
\references{
D. Jaitin, E. Kenigsberg, et al.: Massively parallel single-cell RNA-seq for marker-free decomposition of tissues into cell types. *Science* 343: 776-779
}
