% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{three_dimension_ex_with_t}
\alias{three_dimension_ex_with_t}
\title{Example of The artificial data}
\format{
"" \code{three_dimension_ex_with_t}
A data frame with 600 rows and 4 columns.
The 1st to 3rd columns are the xyz-coordinates of the data.
The 4th column is the corresponding covariate.
}
\usage{
three_dimension_ex_with_t
}
\description{
An Example of the artificial data with t for 3d plot used in Yamazoe and Naito (2023).
}
\keyword{datasets}
