% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{transfer_annotation}
\alias{transfer_annotation}
\title{Function that takes the annotation from one dataset and it transfers it to another one.}
\usage{
transfer_annotation(
  referece_obj,
  query_obj,
  pipeline,
  annotation_col,
  pcDims = 30,
  projectUMAP = FALSE,
  isVerbose = FALSE
)
}
\arguments{
\item{query_obj}{Query seurat object.}

\item{pipeline}{Normalisation method to use (SCT or LogNorm)}

\item{annotation_col}{Name of the column in the metadata slot of the reference object containing the labels to transfer.}

\item{pcDims}{Number of PCs (30 by default)}

\item{projectUMAP}{| Default \code{FALSE}}

\item{isVerbose}{Show all the messages to screen | Default \code{FALSE}}

\item{reference_obj}{Annotated seurat object to use as the reference.}
}
\description{
Function that takes the annotation from one dataset and it transfers it to another one.
}
\examples{
\dontrun{transfer_annotation()}
}
\keyword{seurat}
