% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_utils.R
\name{validate_markers}
\alias{validate_markers}
\title{Validate markers}
\usage{
validate_markers(object, marker_file_path, organism = "human")
}
\arguments{
\item{object}{Seurat object.}

\item{marker_file_path}{output directory}

\item{noFiltering}{FALSE}
}
\description{
Validate markers
}
\examples{
\dontrun{
validate_markers()
}
}
\keyword{markers}
