% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{VlnPlot_modified}
\alias{VlnPlot_modified}
\title{Improved violin plot}
\usage{
VlnPlot_modified(
  srtObject,
  features,
  group_by,
  split_by = NULL,
  idents = NULL,
  invert = FALSE,
  assay,
  slot,
  log = FALSE,
  boxplot = FALSE,
  group_comparison = NULL,
  alpha = 0.6,
  colors_pallet = "viridis",
  title,
  right_labels_col = NULL,
  dup_x_axis = FALSE,
  y_scale = "free_y"
)
}
\arguments{
\item{srtObject}{Seurat object}

\item{features}{Vector of features to plot}

\item{group_by}{Name of one or more metadata columns to group cells}

\item{split_by}{Name of a metadata column to split plot by}

\item{idents}{Identity classes to include in plot (default is all)}

\item{invert}{Invert the Identity selected (idents)}

\item{assay}{Name of assay to use, defaults to the active assay}

\item{slot}{Name of the slot to use}

\item{boxplot}{Overlay a boxplot}

\item{group_comparison}{vector of colours to use to colour the feature names}

\item{colors_pallet}{Color palette to use. viridis by default}

\item{title}{Figure title}
}
\description{
Improved violin plot
}
\examples{
\donttest{
library(Seurat)

features <- c("MS4A1","CD3E","CD8A")
slot <- "data"
assay <- "RNA"
group_by <- "groups"
group_comparison  <- list(c("g1","g2"))
VlnPlot_modified(pbmc_small, features, group_by, split_by = NULL, idents = NULL, 
                  invert = FALSE, assay, slot, boxplot = TRUE, group_comparison = group_comparison,"a title")
}
}
\keyword{seurat}
