% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scR.R
\name{risk_bounds}
\alias{risk_bounds}
\title{Utility function to generate data points for estimation of the VC Dimension of a user-specified binary classification algorithm given a specified sample size.}
\usage{
risk_bounds(x, l, m, model, predictfn, ...)
}
\arguments{
\item{x}{An integer giving the desired sample size for which the target function is to be approximated.}

\item{l}{A positive integer giving dimension (number of input features) of the model.}

\item{m}{A positive integer giving the number of simulations to be performed at each design point (sample size value). Higher values give more accurate results but increase computation time.}

\item{model}{A binary classification model supplied by the user. Must take arguments \code{formula} and \code{data}}

\item{predictfn}{An optional user-defined function giving a custom predict method. If also using a user-defined model, the \code{model} should output an object of class \code{"svrclass"} to avoid errors.}

\item{...}{Additional model parameters to be specified by the user.}
}
\value{
A real number giving the estimated value of Xi(n), the bounding function
}
\description{
Utility function to generate data points for estimation of the VC Dimension of a user-specified binary classification algorithm given a specified sample size.
}
