% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scR.R
\name{risk_bounds}
\alias{risk_bounds}
\title{Utility function to generate data points for estimation of the VC Dimension of a user-specified binary classification algorithm given a specified sample size.}
\usage{
risk_bounds(x, ...)
}
\arguments{
\item{x}{An integer giving the desired sample size for which the target function is to be approximated.}

\item{...}{Additional model parameters to be specified by the user.}
}
\value{
A real number giving the estimated value of Xi(n), the bounding function
}
\description{
Utility function to generate data points for estimation of the VC Dimension of a user-specified binary classification algorithm given a specified sample size.
}
\examples{
mylogit <- function(formula, data){
m <- structure(
  glm(formula=formula,data=data,family=binomial(link="logit")),
  class=c("svrclass","glm")  #IMPORTANT - must use the class svrclass to work correctly
)
return(m)
}
mypred <- function(m,newdata){
out <- predict.glm(m,newdata,type="response")
out <- factor(ifelse(out>0.5,1,0),levels=c("0","1")) #Important - must specify levels to account for possibility of all observations being classified into the same class in smaller samples
return(out)
}
predict.svrclass <- mypred
model <- mylogit
risk_bounds(5)
}
