% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{SCR-package}
\alias{SCR-package}
\title{SCR: Sequential Covariate-adjusted Randomization}
\description{
The SCR package provides function of implement of randomization:
}
\section{SCR functions}{

please ?scrr and ?scrrm to view function usage
}

