% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrr.R
\name{SCR}
\alias{SCR}
\title{Implement of SCRR}
\usage{
SCR(covariate, assignment, d = 5, q1 = 0.75, q2 = 0.85)
}
\arguments{
\item{covariate}{A dataframe of patients' covariates which are needed to balance, the first column is patients' unique ID number and the rest of the columns are corresponding patients' covariates.}

\item{assignment}{A dataframe of patients' allocation which has two columns; If you had allocated partial patients, please input them (the first column is patients' unique ID number and the second colunmn is patients' allocation), please pre-set the allocation group number and enter the number in the second column. IF all the patients are not be allocated, please input 'assignment = NA' directly.}

\item{d}{An integer; restriction for sample sizes.}

\item{q}{A number; allocation probability, default = 0.75}
}
\description{
randomize patients into treatment groups by SCRR
}
\examples{
#simulate covariates of patients
library(MASS)
p=6;n=30
sigma<-diag(p);mean<-c(rep(0,p))
data<- mvrnorm(n, mean, sigma)
covariate<-as.data.frame(data)
#IF all the patients are not be allocated
scrr(covariate = covariate,assignment = NA,d=5,q1=0.75,q2=0.85)
#IF you had allocated partial patients
assignment<-c(1,2,2)
scrr(covariate = covariate,assignment = assignment,d=3,q1=0.75,q2=0.9)
}
