% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot3d.R
\name{scr3d}
\alias{scr3d}
\title{S.C.R for 3d Data}
\usage{
scr3d(
  data,
  t = NULL,
  div = 16,
  alpha = 0.05,
  h_hat = "AM",
  cent = TRUE,
  p = 2,
  grid_points = 201L,
  point_size = 1,
  point_color = "black",
  curve_size = 4,
  curve_color = "black",
  scr_alpha = 0.4,
  scr_color = "#afafaf",
  xlabel = "y1",
  ylabel = "y2",
  zlabel = "y3",
  aspect = TRUE
)
}
\arguments{
\item{data}{Matrix or data.frame with three columns.}

\item{t}{Vector whose length is the same as the number of rows in the data. It represents the covariate. This is not always necessary.}

\item{div}{Numeric. The number of vertices in the cross-section of the region.}

\item{alpha}{Numeric. 1 - confidence coefficient.}

\item{h_hat}{"MAX", "AM", "GM" or "MIN". This parameter determines the choice of h in the Algorithms.}

\item{cent}{Boolean. TRUE, Algorithm2; FALSE, Algorithm1.}

\item{p}{Numeric. This value is used in Algorithm2. The recommend value is written in the paper.}

\item{grid_points}{Numeric. the number of evaluate points.}

\item{point_size}{Numeric. Dot size of the data points in 3d plot.}

\item{point_color}{String. Dot color of the data points in 3d plot.}

\item{curve_size}{Numeric. Width of the estimated curve in 3d plot.}

\item{curve_color}{String. Color of the estimated curve in 3d plot.}

\item{scr_alpha}{Numeric from 0 to 1. Transparency of the SCR in 3d plot.}

\item{scr_color}{String. The color of the SCR in 3d plot.}

\item{xlabel}{String. The label of x-axis.}

\item{ylabel}{String. The label of y-axis.}

\item{zlabel}{String. The label of z-axis.}

\item{aspect}{Boolean. If TRUE (recommended), the figure is drawn to match the actual aspect.}
}
\value{
The figure of SCR and the radius of the cross-section of the region.
}
\description{
Construct S.C.R for 3d Data with Algorithms in the paper.
}
