% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{setup_sc_project}
\alias{setup_sc_project}
\title{Setup project directory structure}
\usage{
setup_sc_project(project_location = getwd(), project_name = "new_project")
}
\arguments{
\item{project_location}{Directory in which the project will be located.  Default is "getwd()"}

\item{project_name}{Name of the project}
}
\value{
no return value.  Creates system folders.
}
\description{
Create reproducible project directory organization when initiating a new analysis.
}
\examples{
\dontrun{
# If using built-in directory structure.

project_dir <- "/home/projects"
projectID <- "new_project"

setup_sc_project(project_dir, projectID)
}
This will create a folder with the following structure

new_project
|--- 01_metadata 
|--- 02_raw_data
|--- 03_processed_data
  |--- cellranger_summaries
  |--- cloupe_files
  |--- Samples_QC
|--- 04_additional_data
|--- 05_scripts
|--- 06_reports
|--- analysis

}
\concept{organization_util}
