% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart.R
\name{smart}
\alias{smart}
\title{Implement of SMART}
\usage{
smart(covariate, assignment, K, d, q = 0.75, method = "mean")
}
\arguments{
\item{covariate}{A dataframe of patients' covariates which are needed to balance, the first column is patients' unique ID number and the rest of the columns are corresponding patients' covariates.}

\item{assignment}{A dataframe of patients' allocation which has two columns; If you had allocated partial patients, please input them (the first column is patients' unique ID number and the second colunmn is patients' allocation), please pre-set the allocation group number and enter the number in the second column. IF all the patients are not be allocated, please input 'assignment = NA' directly.}

\item{K}{An integer; number of assignments of the trial.}

\item{d}{An integer; restriction for sample sizes.}

\item{q}{A number; allocation probability, default = 0.75}

\item{method}{If K=2, please IGNORE this parameter and don't input anything; If K>2, you can input one of these texts: 'mean', 'max' or 'median'.}
}
\description{
randomize patients into treatment groups by SMART
}
\examples{
#simulate covariates of patients
p=6;n=30
sigma<-diag(p);mean<-c(rep(0,p))
data <- mvrnorm(n, mean, sigma)
covariate<-as.data.frame(data)
#IF all the patients are not be allocated
smart(covariate = covariate,assignment = NA,K=3,d=5,q=0.75)
#IF you had allocated partial patients
assignment<-c(1,2,2)
smart(covariate = covariate,assignment = assignment,K=3,d=3,q=0.75)
}
